fix(extensions): moved yfm file widget and paste to yfm preset
