+++ src/extensions/yfm/YfmFile/YfmFilePaste/skeleton.scss
