+++ src/extensions/yfm/YfmFile/YfmFilePaste/upload.ts
@@ -1,9 +1,9 @@
 import {Node} from 'prosemirror-model';
 import {EditorView} from 'prosemirror-view';
 
-import {fileType} from '../../../../extensions/yfm';
 import {FileUploadHandler, UploadSuccessItem} from '../../../../utils/upload';
-import {FilesBatchUploadProcess} from '../../utils/upload';
+import {FilesBatchUploadProcess} from '../../../behavior/utils/upload';
+import {fileType} from '../YfmFileSpecs';
 import {createFileNode} from '../utils';
 
 import {FileSkeletonDescriptor} from './skeleton';
