+++ src/extensions/yfm/YfmFile/YfmFileWidget/view.scss
